add setting read/write methods to MZIMEJA

+++ convert.cpp
@@ -186,7 +186,7 @@ BOOL MZIMEJA::DeployDictData(
         *pch++ = L'\n';
       }
       *pch++ = L'\0'; // NUL
-      assert(size == pch - reinterpret_cast<WCHAR *>(pv));
+      assert(size == size_t(pch - reinterpret_cast<WCHAR *>(pv)));
       ::UnmapViewOfFile(pv);
       ret = TRUE; // success
     }
